pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grab"
version = "1.0.0"
description = "YouTube & Instagram content grabber — Obsidian-ready markdown with embedded media"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.0.0",
    "playwright>=1.40",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.scripts]
grab = "grab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/grab"]
